refactor: check token existence

+++ client/app/src/services/adminService.js
@@ -1,9 +1,16 @@
 import axios from "axios";
 import { API_ENDPOINTS } from "./api";
+import authService from "@/services/authService";
 
 const adminService = {
   
-  async getAllUsers(token) {
+  async getAllUsers() {
+    const token = authService.getUserToken();
+    
+    if (!token) {
+      return false;
+    }
+
     try {
 
       const response = await axios.get(
@@ -24,7 +31,13 @@ const adminService = {
     console.log(error);
   }
   },
-  async updateUserRoles(userId, newRoles, token) {
+  async updateUserRoles(userId, newRoles) {
+    const token = authService.getUserToken();
+    
+    if (!token) {
+      return false;
+    }
+
     try {
       const response = await axios.put(
         API_ENDPOINTS.USERS.ROLES(userId), {
@@ -43,7 +56,13 @@ const adminService = {
       console.log(error);
     }
   },
-  async deleteUserAsAdmin(token) {
+  async deleteUserAsAdmin(userId) {
+    const token = authService.getUserToken();
+    
+    if (!token) {
+      return false;
+    }
+    
     try {
       const response = await axios.delete(
        API_ENDPOINTS.USERS.BY_ID(userId),
